Move observation helpers out of specs base module.
Keep base.py limited to ActionSpec/ObservationSpec ABCs by relocating mean_clean_area and hours_normalized to PlantGrowthChamber utils.
