+++ src/environments/PlantGrowthChamber/specs/base.py
@@ -33,14 +33,3 @@ class ActionSpec(ABC):
     @abstractmethod
     def decode(self, action: Any, backend: Any) -> np.ndarray:
         pass
-
-
-def mean_clean_area(df: pd.DataFrame) -> float:
-    if df.empty or "clean_area" not in df.columns:
-        return 0.0
-    return float(df["clean_area"].mean())
-
-
-def hours_normalized(local_time: datetime) -> float:
-    hours_since_start = (local_time.hour - 9) + ((local_time.minute - 30) / 60)
-    return float(np.clip(hours_since_start / 11.0, 0, 1))
